Sentry: ignore some specific errors
We've tried to get to the bottom of these errors in sentry and can't.

They certainly tend to occur more on some devices that others, but we
still can't figure out why. We also have no reports of actual issues
from users, just error records in Sentry.

So, let's stop seeing this noise and move on with our lives.