fix(cors): use `x-ifixit-forwarded-host`
It seems that Vercel rewrites the `x-forwarded-host`, so we can't tell if the app is being proxied in our dev environment.

In https://github.com/iFixit/ifixit/pull/45558 we added this custom header so that we can detect if the app is being proxied in dev.

Also, pass the context into the host-detection callback so the caller doesn't have to drill into the context and do type coercion.