Pro Users: Add test
This verifies that pro users get a discount on products. We are using
the repair business toolkit and mocking the logged in user to verify
that they get the discount.

Because we are not mocking the product itself, I didn't want to get
too specific with prices. Instead, we just verify that the price
for a regular user is *higher* than it is for a pro user.