Youtube: add react-lite-yt-embed dependency
Youtube iframes slow down a page a suprising amount, many requests, 1MB
of resources, slow JS, ...

A way to combat this is to show the thumbnail and the play button and
only create the iframe when the play button or thumbnail is clicked.

This is the most depended upon package that does this. It's only 3KB
minified+gzipped.