Product Videos: use a lite version of the YT embed
Now that we have structured data available, we can use it to
conditionally embed a lite version of the YT player.

This lite version effectively just shows the thumbnail wrapped up to
look like the youtube iframe and then swaps it for the iframe when the
user clicks. This makes a huge difference in page speed cause the YT
player is heavy weight.