ProductVideos: Stylize and config lite-youtube
The patching is quite annoying, took me a while to find a config that
worked. Effectively, their default index.js in dist/ does a `require()`
of a file that does `import React ...`

Next.js fails to compile with an error:

> SyntaxError: Cannot use import statement outside a module

I imagine there are several ways to deal with this, I tried several and
this was the one that worked with the fewest downsides.

Attempted solutions
====

* Import asbolute path of CJS module file
  * Worked!
  * Typescript couldn't find the index.d.ts file anymore cause I wasn't
    importing the main pacakge, but a particular file
* Converting the main frontend/package.json to "type":"module"
  * Mostly had to convert all the next xonfig to use `import` and
    `export`
  * Didn't actually solve the problem with the react-lite-yt-embed
    package
* Changing some options in tsconfig
  * Didn't seem to work

Otherwise, the change works great, there are many fewer requests and
less data downloaded until clicking the video.