View Stats Tracking: add view count pixel
Our old product page used a tracking pixel from our internal pixel-ping
service. This only counts views in aggregate, no PII is kept or cookies
set. this is the same service and stats used on other doc types.

We forgot to include this when re-creating the product page because
there's no visible component.