fix(xsell): cards should be links
Makes the cross sell product cards links. Clicking the checkbox will select/deselect the cross sell product.

Note that `onKeyUp` was changed to `onKeyDown`. Without this change, clicking "Enter" with the checkbox focused caused the link to be followed (and the checkbox to be toggled). The expected behavior is to toggle the checkbox and not follow the link, and using `onKeyDown` achieves that.

Next link docs: https://nextjs.org/docs/api-reference/next/link#if-the-child-is-a-custom-component-that-wraps-an-a-tag