fix: hide children that link to empty collections
Hides product list children links that link to collections that are empty.

Intersects the available device facets with the current product list children to determine which children are non-empty. This will not work for pages that have >200 device facets. If that proves to be a problem, we can address it later on.