Footer: defer to client-side
The footer is a lot of html nearly (40KB) and since it's usually far
below the fold, let's defer it to rendering on the client side.

To support this pattern I've added a hook that should help:

* Returns false on the server side
* Returns false during hydration on the client
  * React requires the first client render produce the exact same DOM
    state during hydration
* Re-renders right after hydration
* Returns true on every further re-render/re-mount, without returning false before `useEffect()` runs

Note: the footer is only mounted once, but this will be useful for other
components that are unmounted when switching products.