modified transformation rules for jest
We are having some issues trying to transform some
TSX imports because the packages use `ESM Modules` and Jest does not natively support
`ESM Modules`. As result, we get an `unexpected token 'export'` error when trying to render some
of our React components. By setting an empty array for `transformIgnorePattern` it will override the default of not transforiming `/node_modules` packages.

In addition, we also have some `css` module imports that Jest cannot handle with the `identitiy-obj-proxy` package and therefore need an additional transformer package.

https://github.com/facebook/jest/issues/3094#issuecomment-514308251