use next/jest to simplify Jest configuration
Next.js has added a wrapper around configuring Jest to run tests. This
wrapper will do a lot of things we are already doing in our configuration.

The ones we are more paticularly interested in is the auto-mocking of
`.css` modules and its variants, as well as mocking image imports. This
means we can remove our own file mock and configuration settings. As well
as the default test paths that get ignored.

In addition, `next/jest` uses `SWC` to transform the files which is a
lot faster than Babel and is something we have already implemented in
our monorepo
https://github.com/iFixit/ifixit/commit/37985f37d022fe8ddfc3ac822f953253a413933e.

Lastly, because of the default settings for Jest that `next/jest` has
we needed to manually override the `transformIgnorePatterns` to use
an empty array and thus whitelist all node_modules paths.

https://github.com/piglovesyou/graphql-let/issues/610#issuecomment-981570694

All the default configurations can be found here:
https://github.com/vercel/next.js/blob/canary/packages/next/build/jest/jest.ts#L101C12-L179