create test util function to render components with context
We can no longer simply mock @ifixit/app and therefore need to
wrap the components we want to test with `AppContext` in order to render
them for testing. Instead of appending this component to each location
we use render, let's create a utility function that will render
the component(s) we pass with the `AppContext` wrapped around it.