reuse WikiHtmlAccordion for product specifications
If we move the styling for the product specifications to the
WikiHtmlAccordion component, then we can reuse the same
component to render the specifications section similar to
the way render the Kit Contents section and the Assembly Contents
section. Therefore, we remove duplicate code and reduce the complexity
of product section.