move zoomed product image component to top of component tree
At first glance, this component looks like it is part of the list
of components that give information about the product, but it is
actually the container for displaying the zoomed-in image of the
product when hovering over the main product image. By moving this
component to the top of the component tree, we can make it easier
to understand the structure of the product page and not misassociate it
as part of the product information.