add package.json script to run playwright tests
Similar to the Cypress script, we can now run the Playwright tests with
`pnpm playwright:run`. In addition, we are making use of the webServer
option in the Playwright config to automatically kick off the dev server
if it is not already running. This is a lot more convienent than
Cypress that requires the dev server to be running before running the
tests.