extend jest test utils
While trying to use the renderWithProviders function to render the
Product Section component with AppContext, it kept failing trying to
use the `useBreakpoint` hook from Chakra. This was due to us lacking
the Chakra provider when trying to render the components. Instead of
adding just the Chakra provider as another wrapper to the renderWithContext
function, I decided to import the AppProviders component that includes
all the providers we need to render our components in the main app.

After resolving this issue I ran into another issue where the Jest would
fail with the error `TypeError: window.matchMedia is not a function`. This
was due to some of our code trying to use a method which the JSDOM
has not implemented yet. This is a common issue and Jest has written a
guide on how to resolve this issue. The guide can be found here:
https://jestjs.io/docs/manual-mocks#mocking-methods-which-are-not-implemented-in-jsdom