create utility functions to mock product data for jest
This mimics the API we have in our ProductExtensions class within
the iFixit/ifixit repo. This will allow us to easily get a Product
like object for testing purposes and give us the ability to override
any of the properties we want to test. In addition, since we are using
the `Partial` type, we can make the properties optional but still
retain the type safety when trying to override default values.