convert add_to_cart cypress syntax to playwright syntax
In order to get this test to pass for Playwright, I had to make
a few modifications to the Playwright config file. I needed to increase
the expect timeout time and I needed to slow down the browser operations.
The need to slow down the browser operations was due to the
Playwright speeding through the interactions and our app not being
able to keep up with the speed, which resulted in the test failing
from the app not being in the correct state.