CompatibleDevice: Add tests
This is a basic test to ensure that the base component renders
correctly as well as the component renders correctly when the
number of variants exceeds the limit. In that case, the component
should display the `And 3 other models...` text. In this case,
I wanted to generalize it a bit more so I only included
`other models...` in the query text.