port collections_scroll to playwright
In order to port this test I needed to extend the expect
functionality to support Cypress's `isWithinViewport` assertion.
Playwright's `isVisible` assertion simply checks if the element is
is in the DOM and not hidden. In order to check if an element is
within the viewport, I needed to use the `boundingBox` property
of the element. This returns the element's position and size
relative to the viewport. Therefore, using the page viewport size
we can check if the element is within the viewport.

I also needed to add an additional wait before validating the
element is within the viewport. This is because we are still
in the process of scrolling to the top of the page when we
are validating the element is within the viewport. This will
wait until all network requests have finished before validating
the element is within the viewport.