port parts_page_devices test to playwright
I refactored the test a bit by moving the code from the helper functions
into the main test as it was only used once. In addition, I added
an additional wait as when we click the child link the page has not
fully loaded yet and as a result the page.url() call returns the
previously loaded page url which causes the test to fail.