port filters test to playwright
This test was a little tricky to port because Playwright's locators
do not bind the element of the pages. This means everytime we
want to interact with a Locator, it will requery the page with the
specified selector. If the element has changed such as when we expand
a collapsed facet, then the Locator will not be able to find it anymore
when we try to continue interacting with it. To solve this, we need to
use the actual ElementHandle that the Locator returns and use that to
interact with the elements.

Additionally, in Cypress we have a custom command to append a
`filteredProducts` property to the window object in
https://github.com/iFixit/react-commerce/blob/main/frontend/templates/product-list/sections/FilterableProductsSection/index.tsx#L56.
This property is later used to validate that products we got back
from the `search` query include the same `facet tag` as the one we
selected in the filter. In order to test this, we need to return
the response from the `search` query and then use that to validate the
products include the correct `facet tag` as the one we selected. There
was an additional problem with this as Playwright does not have an
`its` command like Cypress does. To resolve this, I added a utility
function that wraps around `lodash.get()` to get the nested property
value of an object with a string path.