use manual timeout instead of network state
Waiting for network idle is not reliable enough, and there is no other
reliable way to know when the page has scrolled to the top of the page
without validating the element is within the viewport, which is what
this test is about. Therefore, we will manually wait for 1 second
before asserting that the search box is wthin the viewport.