update playwright config to support CI
Let's set the number of parallel worker, in CI, equal to 100% of available
cores. This should help speed up the tests in CI.

Additionally, let's simplify the reporter to just a list when
running the tests as we don't need to spend extra time generating
the HTML report.

Lastly, we need to make some modificaitons to the webserver config. If
we are running in CI, then we don't need to start the backend server
so we only need to run the frontend server. Additionally, we need to
build the frontend application similar to production so we will
need to run the commands `pnpm build && pnpm start` for CI but not
for local development.

Co-authored-by: ardelato <angel.dlt.1996@gmail.com>