glob api request for newsletter subscribe
We were using a path of local host to intercept the api request instead
of globbing all api requests to the newsletter subscribe endpoint.
This will catch the api request if sent to cominor. This is the same
syntax we are using for the other api requests, which are currently
passing.

Co-authored-by: Bat Battur <bat@ifixit.com>