add msw servers and handlers for playwright tests
This just adds the base files necessary to set up the msw server, worker,
and handlers for the playwright tests. The handlers are not yet used in
the tests, but they will be in a future commit.

In addition, we will only setup the msw server as there is a compatibility
issue between `Playwright`'s `page.route` and `msw`'s interceptors. Until,
we can dynamically modify the handlers, we will only setup the mocks
for the server.