conditionally intercept findProduct request
We are still in the process of trying to figure out how to dynamically
add handlers. However, that will block us from moving forward and
writing more tests. So, we are going to conditionally intercept the
request based on the product handle we are looking for. For now, if
we are looking for the product with the handle 'caig-electronic-cleaner',
which is the same product we currently have mocked in our mocked data,
then we will intercept the request and return the mocked data. Otherwise,
we will request the original data and return it.