use shopify's quantity instead of cart api's
We currently use two different apis to get the quantity of a product.
We use shopify's api to get the starting quantity listed on the product
sectio. However, once we add the product to the cart, we use the cart api
to get the quantity for the product section. As a result, this causes
the mocking of product to be useless after adding the product to the cart.
Instead of also mocking the cart api, we will just default to using
the shopify api for the quantity first then the cart api if the value
is not available.