add separate handlers for stock level of product
Instead of using the same handler for the mocked product, let's use
a separate handler to distinguish between the stock level for the mocked
product query. This way we do not need to create a separate mocked
product query for each stock level nor will this interfere with the
original product handle.

In addition, let's lower the stock level so we do not need to add
a bunch of items to the cart to run out of stock.