update: out of stock product test
If a product variant is out of stock and the product has another variant
that is in stock, then we will automatically select the in stock variant.
Therefore we need to update the test to reflect this. To note, the
variant selectors used are what we use in the `fix_kit` test.

In addition, we will now check that the in stock variant will work as
expected.