Add test for a product with multiple images
This test grabs the src of the currently displayed image on the
product page using the helper function, then clicks on the next image.

Then when the next image becomes the active image, it also grabs the
src and compares it with the first image src to make sure different
images are displayed.