add mobile devices for playwright testing
These devices come from a preset list of devices that Playwright provides.
https://github.com/microsoft/playwright/blob/main/packages/playwright-core/src/server/deviceDescriptorsSource.json

They simply emulate the device's screen size and user agent. They do not
actually emulate the device's hardware.

To note: we still need to figure out how to enable cross-site cookies
for the `webkit` browser in order to succesfully run the tests on
Safari. Therefore, we will not be testing on `iPhone` or `iPad` devices
for now.