Add breadcrumb test for mobile and table viewport width size
On viewport width sizes below 1000px, we display only the last child breadcrumb
regardless of the number of breadcrumb links. Everything else is collapsed in
the breadcrumb menu.

The test asserts that the last child breadcrumb and the mobile breadcrumb menu
are visible.

Then it asserts that ancestor breadcrumb links are not visible on mobile.