remove mention of Safari project
It doesn't look like there is a straightforward way to disable CORS for
webkit. As opposed to Chromium, there are no launch option argument flags
that we can pass to WebKit to disable the security features. I also tried
intercepting all requests and removing the `sec-fetch` headers
as a way to fake a no-cors request but that didn't seem to work
either.

Therefore, we will not be able to run the tests in WebKit without some
convulted setup to disable CORS. I don't think it's worth the effort
to do that. We can revisit this if there is a need to run the tests
in WebKit, but for now let's drop the mention of the Safari project
in the config.