Add collections filters mobile test for drawer buttons.
For the mobile facet drawer, the user either can select the filter
and click close button or select the Apply button. For either case,
the search result gets updated.

This test uses the `Apply` and `Clear all` button in the mobile facet
drawer. The previous test used the `X` close button.