update playwright debug script
We made this pnpm script prior to changing some of our playwright project
names - https://github.com/iFixit/react-commerce/pull/1223. As a result,
the script fails because the project name `chromium` no longer exists
in the playwright config. This updates the script to use the new project
name `Desktop Chromium`.