Homepage KPIs: Use required elements
The `dl` element is required to have a `dt` element followed by `dd`
elements (if `div` elements aren't being used).[^1] The Chakra docs
seem to suggest that `StatLabel` is intended to be used for text of
the sort we're rendering in `StatHelpText` right now.[^2] `StatLabel`
renders as a `dt` while `StatNumber` and `StatHelpText` render as
`dd`s; thus, using `StatLabel` would provide the required `dt`
element. Because the docs require that the `dt` appear before the
`dd`, this moves the `StatLabel` to before the `StatNumber` and uses
the `flex-direction` `column-reverse` value to make them appear in the
reverse order.

[^1]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/dl
[^2]: https://chakra-ui.com/docs/components/stat/usage