create a product page fixture
This fixture will be used to make it easier to interact with the product page. For now we will start with a helper function to add a product to the cart, to go to product page via the product handle, and lastly, to grab the product sku from the product page.

We can append more fixtures in the future as needed.