add and update data-testid for produt prices
We target different prices of a product in our tests but only have a single data-testid attached to a price. We need to add more data-testid to the price to target the current price, the original price (ie compare at price), and the discount percentage. This will allow us to target the correct price in our tests. In addition, since we are adding these data-testid, to the price component we need to
add an additional data-testid to prices for the main product and not the featured products or cross-sell items.

That said, I updated all references of `product-price` to `current-price` and updated the snapshot.