add utility functions to parse for the product price
These functions will make it easier to nab the prices (current and original prices) of the product page we are on. This will also assert that the
prices are visible and in the correct format.

That said, let's replace the duplicated code in the tests with the new
utility functions.