add utility functions to make it easier to select a product variant
We have a hack to select a product variant by searching for the assumed
variant text. This is brittle and will break if the variant text changes
or we go to another product page. By including an arial-selected attribute
we can switch between the two variants. In addition, if we add a
data-testid to the parent component, then we can make it easier to
select a variant for other selector types by filter down from the
parent component.