refactor: return a number instead of a string for the price
Let's have the Devs decide what they want to do with the price. Most of
the interactions with the price are done with numbers, so it makes sense
to return a number instead of a string. If the Devs want to work with
strings, they can always convert the number to a string with `toFixed(2)`.