fix: use child device titles in algolia query
Uses the child device titles in the algolia query to filter empty child collections. Fixes a bug where mismatches in the Strapi and iFixit device hierarchy caused some children to be erroneously hidden.

For example, vacuum parts live under `Appliance` in iFixit, and `Electronics` in Strapi. This mismatch caused the `Vacuum Cleaner Parts` child collection to be hidden on the /Parts/Electronics page.