Troubleshooting: Add keys for iterators
I'm not sure exactly how good these are, but I think they should
probably do. I'm using the heading for each problem as the key for
problems; it seems reasonable to assume those headings are unique,
given the logic of problem pages.

`EndMatter` is trickier: there's no reason the headings need to be
unique. (According to the code, the headings could even be `null` for
end matter; I don't believe that can actually happen in practice,
because the PHP would treat any text without a heading at the end of a
wiki as being part of the last heading (naturally, because that's what
it would be). And it appears it's not possible to create a heading in
our wikitext without any heading text. So I think we don't have to
deal with the `null` case, but we still might need to deal with
duplicated headings.) I decided to use the body of the text as the
key; it seems pretty unlikely that we'll have multiple end-matter
sections with duplicated text.