add utility function to get current state of the cart drawer
This will check whether or no the cart drawer is currently 'open' or
'closed' by checking whether the `close` button is visible or not. In
the case the `close` button is visible, then the cart drawer is open.
Otherwise, it is closed.

This is useful in making assertions about the state of the cart drawer.