fix(svg): micron partner badge dimensions
Modifies the width and height of the Crucial and Micron SVG's to be 96x64, like the rest of our partner SVG's. The original aspect ratio (`viewBox`) is preserved.

The Micron and Crucial images do not have horizontal padding, but it seems like the others do, so padding will need to be applied via CSS downstream.