Footer Copyright: remove year
We got 1000s of hydration errors around new-years where
this value is different between server and client.

The year isn't particularly relevant here as copyrights last a long
time. Technically, we don't even need a copyright message, but claims
are more defensible if it was clear your content is copyrighted by you.