move bootstrapping msw server to a callable function
Previously, we only need to bootstrap the msw server once when we were
starting the nextjs server. Therefore, we could just start the server
when the file is imported.

However, we will soon reuse the same msw server for our custom nextjs
server to allow for dynamic handlers. Therefore, we need to be able to
start the server at any time not just when we import the file.