use the new msw fixtures within the tests
We can now move the static handlers into the respective tests as
dynamic handlers. This will allow us to remove the conditionally
set static handlers and be able to write more handlers in the future.

There were a few things that needed to be changed to make this work.

First, since the serverRequestInterceptor, or msw server, was created
on a different port, we needed to update the `goto` method to use the
correct port when navigating to the product page. This is why
we passed the port as a fixture to be used in this manner.

Secondly, we needed to update the test timeout to be longer than the
default 60 seconds since the bootup of the server can take a bit. Instead,
of increasing the timeout for all tests we can set the timeout property
for the worker fixture. This will allow us to set the timeout for all
tests that use the worker fixture, i.e. that use the custom server, without
affecting the timeout on regular tests.

Lastly, we needed to modify the logic for when the customServer fixture
would be instantiated. Previously, it would default to being lazily loaded,
meaning until a test called for it, it would not be instantiated. This is
a problem if we have multiple workers starting up the server at the same.
If two workers are booting up the server at the same time, then there
will be a build conflict as two instances of webpack are building the app
to the same directory. To avoid this, we can set the `auto` property to
true only if we are in CI. This will ensure that the server is always
instantiated before the tests run and therefore avoids the build conflict.
While in development mode, we should only use a single worker to avoid
the build conflict.