remove setting of NEXT_PUBLIC_MOCK_API env within playwright npm script
This is no longer needed as we are now manually booting up the Next.js
server ourselves within the Playwright test suite. This server will
have the MSW layer already integrated without the need for the
env variable.

However, if we are trying to create or troubleshoot MSW handlers, then
being able to enable the MSW layer for the normal Next.js server is
still useful.