move build step to external github action step and cache it
We sometimes failed when running the tests because Playwright would need to build and startup the server first.

By moving the build step to a prior step in the GitHub action workflow, we can ensure that Playwright just needs to start and wait for the server to be ready, as opposed to
building the app and starting the server.