add timeout to playwright github action
There is currently an issue with this action that requires further
investigation. Playwright seems to be unable to communicate with the
`backend` server (i.e. localhost:1337) which is causing the tests to
be re-ran multiple times and eventually fail with a total time of about
30 minutes.

For now let's set a timeout of 15 minutes to prevent the action from
running more than normal.