update comment about starting msw for the browser
We will now be able to dynamically add handlers to msw, so this comment
is partially incorrect. We will will not be using `page.route` in the
future to intercept client-side requests but we will still not use an
`msw` browser side serivce worker to intercept requests -- at least in
Playwright. We will only use the `msw` browser side service worker
when running the dev server with API mocking enabled.