Caching: standardize event and timing logs
For parsing these values with datadog, it's helpful to have a less
common value separator that isn't present in non-timing messages.
To ensure we get a consistent format, add some timing functions to our
logger.

I opted for an explicit stat name instead of implicitly re-using the
endpoint. The endpoint isn't really meant to be a human readable string
while the timing logs *are*.