Caching: include revalidation time under "stale"
Hitting a stale entry always results in us firing off the "revalidation"
request and waiting some arbitrary amount of time. Since this time is
necessary, let's not pretend it doesn't exist. Stale entries still take
time to deliver the results.

Also, for timing the "set" operation, only include the time to talk to
redis. The "miss" timing already includes the time spent fetching a
fresh value, so let's not include that time in "set" as well.