move international buybox request to handlers file
The API request `api/2.0/internal/international_store_promotion/buybox`
is used to determine if the user is in Europe or the US. If they are in
Europe then a different component will be rendered instead of the
`Add to Cart` button. This is a problem for the European Devs since they
will need to use a VPN and whitelist their new IP to the Dev Server to
get the tests to pass. That is why this request was originally mocked.

However, we would need to mock this request for all tests to ensure
we would be running the tests in the "US". As such, I moved this request
into the main handlers file to be used by default.