properly wait for the mock service workers
It seems the browser-side service worker finished starting after the
page has already loaded. As such, the requests were not mocked properly.

In order to wait for the browser-side service worker to finish, I
followed similar steps as seen in this blog post by Dillon Shook
https://dillonshook.com/testing-nextjs-with-cypress-and-msw/. We will
create a state to indicate whether the service worker is ready or not.
Then we will return a loading page until the service worker is ready.
After which, we will render the page as normal.