Breadcrumbs: Update to match existing code
This updates the new breadcrumbs implementation copied from iFixit to
be nearly compatible with the existing one in react-commerce. It
includes a number of changes to the breadcrumbs code to make it
type-compatible with the previous implementation, along with a number
of changes because the type-checking in this repo is stricter than in
the source repo.

It also includes a change to the icon library we're using: the current
breadcrumbs code in this repo uses `pro-solid-svg-icons`, so let's use
those here as well.

This does not start using the new implementation; we tried that, but
found that we were ending up introducing an extra ~7kb of JS, but
couldn't tell where it was coming from. Rather than solve that right
now, we've decided to just use this on the new Troubleshooting pages.