reduce number of workers for local runs
GitHub Actions are set to 2 workers due to the two cores available.
However, when running locally there are more cores available as such
we could potentially have 4-8 workers running at the same time. This
causes conflicts and increases the chance of timeouts. Let's reduce
the number of workers to a quarter of the available cores.