allow using a different dist dir for next build
One of the reasons why the playwright tests were timing out was because
each worker was building the custom next server to the same directory.
As such files were being overwritten while the tests were running.

By using a different build directory for each worker we can avoid this
problem. Although, the only way to do this was to set a process env
variable and then use that in the next config. This is a bit of a hack
but it is necessary because the next config is a module and not a
a JSON file so we can't import the config and override the distDir
property in the custom-nextjs-server file.