Add the next-plugin-bundle-stats plugin to be run on `ANALYZE==true`
This is the other plugin we run on iFixit/ifixit; I wanted to see if
it would provide better information for figuring out what's going on
with a build. It wasn't able to answer the question I had, but I
figure it seems worth keeping the set up work I did.