Cache: use null adapter while redis isn't ready
We seem to be getting about ~250 errors per day that say the redis
connection isn't a writable stream. These should be able to be caught
and dealt wth by returning as a miss, but they seem to be aborting the
request.

Redis is communicated with via a persistent connnection, so here we
change caching to dynamically choose the backend based on the connection
state. If it's anything other than "ready", we use the null adapter
(everything looks like a miss).

Possible connection states from Typescript:

type RedisStatus = "wait"
  | "reconnecting"
  | "connecting"
  | "connect"
  | "ready"
  | "close"
  | "end"

Looking at the client code, checking for `ready` seems like the right
choice. I've also upped the connection timeout now that we have a
fallback while connecting and that connecting won't delay any requests.