define methods types for each request to check API type
Since we are using the same request handler for all request types,
we need to define the types for each request type to check the API type.
This will help us in determining the type of handler to create and
return.

That said, the methods availabe for each request are only a subset of
all methods available for each api type. This is because we are using
`msw` to mock the requests and `msw` only supports a subset of all
methods available for each api type.

https://mswjs.io/docs/api/graphql#operation-kind
https://mswjs.io/docs/api/rest#methods

Lastly, in order to check the API type, we will need to check which
method of the subsets availabe is being used. For example, if the
method is one of GraphQLMethods then the API type is GraphQL. Then
by using a predicate return type we can ensure that Typescript
safe guards the methods availabe for each API type and helps with type
narrowing.