define a type for the mocked response options
To create a handler for a request, we also need to know in what way we
should respond. For all the options available, `MSW` provides a `Context`
interface to modify the response instead of modifying it directly
https://mswjs.io/docs/api/response#raw-usage. As such, I had to consider
how to abstract this away from the user to make it easier to use, as
there are multiple methods to transforming the response body and
modifying other response options.

I decided to make the `body` type a union of all the possible types
we can return and to include a `responseType` property to specify
how we should transform the body. By doing this, we can avoid having to
create a subset of helper commands based on all `Context` transform
methods. To note, the `Context` transformer methods will also set the
appropriate `Content-Type` header for the response.