return each type of request handler
Now that we have the minimal required parameters, we can create the
request handlers for each type of request.

These handler classes are not in the `MSW` documentation and were found
in the source code of the library. They are already exported so we can
use them without having to modify the `msw` library. Both of these
handlers implement the abstract `RequestHandler` class which means
we still still use the `RequestHandler` type for the return type.

The main reason for using these handler classes is it made it easier
to abstract the creation of the handlers and reduce the need to create
a bunch of helper functions. If we were to follow the documentation,
then we would need to import `rest` and `graphql` directly
and then determine which method to call while retaining type checks.
An example of this can be seen in:
https://github.com/iFixit/react-commerce/blob/85aaa8ea8e04973cb77bc0627f95754df61c9e25/frontend/tests/playwright/fixtures/request-handler.ts

Lastly a couple things to note:
- The `GraphQLHandler` maps our `endpoint` to their `operationName: GraphQLHandlerNameSelector` parameter which is a bit confusing.
- The `GraphQLHandler` also requires an `endpoint: Path` parameter which
   is the path of the graphql endpoint, basically the location of the
   request. In our case, we don't care about the location of the request
   so we are using the `*` wildcard to match any path. This was apparently
   how the `msw` library was setting this parameter when using the
   `graphql.query` method. I found this out by adding a `console.log`
   in the index file of the `msw` library within our `node_modules`.
- The `responseType` `raw` option is not an actual transformer method.
   The actual transformer method is `body` which can be confusing
   since `body` is used in other locations as well. By using `raw` we
   can inform the user what response type they will actually be without
   creating a burden on the user to decipher the context of which `body`
   we are referring to.
   @see https://mswjs.io/docs/api/context/body