include once and passtrhough options to the handler
These options allow us to configure the handler.

The `once` option allows us to have a one time handler. This means once
the handler is called, it will be marked as resolved and all subsequent
request matches to this handler are ignored, and fall-through to the next
suitable handler.
https://mswjs.io/docs/api/response/once

The `passthrough` option allows us to have a handler that will
actually send the request without it being intercepted by msw.

These options are useful for testing purposes, where we want to test
to override some specific handlers without having to reset them
after each test.