overload the create method
By overloading the create method, we can define individual signatures
for each handler type. This allows us to constrain the parameter types
to the correct ones for each handler type.

For example, if we want to create a Grahpql handler, we need to ensure
that the body can only be an object. By overloading the create method,
we can have type safety without duplicating a bunch of code.